Add manifest for building as a flatpak
It's a bit rough but it includes a desktop file and everything needed 
for the emulator to run.
It enables pulseaudio for sound, network for wifi emulation, device=all 
for gamepad support and filesystem=home to enable the emulator to write 
the save files to the same folder as the rom.

Because melonds uses sdl2 and gtk3, it should run on wayland, but I'm 
not including the socket because I haven't personally tested it to work.