Update CRC32 table to compile-time generation
This is an update to the CRC32 implementation to use a compile-time
generated CRC32 table rather than a run-time generated table protected
with a boolean.

Using `constexpr`, the table is guaranteed to be created statically using
C++17 features(will re-base once https://github.com/Arisotura/melonDS/pull/990 gets merged in).

Verified that the tables are statically generated across GCC, Clang, and MSVC:
https://godbolt.org/z/75Ta77

Verified CRCs are consistent with the previous implementation

This is based on code from another project of mines to reduce runtime load.
https://github.com/Wunkolo/qCheck