add ability to set time, exposed in qt ui too
This approach differs slightly from Suuper's approach. RTC can be offset by a basetime. basetime = 0 means real time is used on boot, but it can still be changed in the firmware.
The passage of time is ultimately based on real time, although API is set so this real time is frontend defined, so BizHawk can set it to its own RTC implementation.