rewrites UpdateClipMatrix to avoid 2 memcpies
Small optimization of UpdateClipMatrix. There is of course the
possibility that at O3 the compiler can tell those copies are
superfluous and optimize them away anyway. Although I haven't done super
rigorous tests, comparing the uncapped framerates of versions compiled
with and without the change the new version does seem to be generally a
bit faster depending on the scene. On the startup screen of Pokemon
Black the new version seemed to be about a couple to a few% faster than
the old one, but about the same on starter selection.
MatrixMult4x4 was also rewritten for better reuse, but unfortunately
it's no longer in style with the other matrix multiplication functions.

Fun observation, because there are 53 bits of mantissa in a double and
64-12=52<53, it's possible to convert these integers to doubles, perform
the calculation on them and then convert back again without losing any
precision. This may sound completely goofy, but compilers are better at
vectorizing floating point arithmetic as you can see on e.g. compiler
explorer, so such a thing could possibly improve performance in some
scenarios. I tried it, and alas, even though the improved vectorization
from using floating points made a good effort, it just wasn't enough to
make up for the extra copies and conversions. I even tried using BLAS,
but that didn't do any better. But that does raise an interesting
thought, if the compilers don't vectorize the integer instructions that
much, could the matrix multiplications be sped up with explicit use of
SIMD intrinsics? Possibly with the vectorclass library? Another thing I
haven't tried is getting rid of the memcpy in the
multiplications. This would probably require heap allocating the arrays
so that pointers could be reassigned, and that extra heap faffery would
probably more than negate the improvement from getting rid of the memcpy.
If you add some intermediate arrays maybe there's some way to shuffle
stack allocated arrays around to get around heap allocating.

Another thing I tried was multithreading, although I only tried the
standard threading library. To begin with, since these matrices aren't
that big I suspected the overhead from threading would negate or
overwhelm any possible speedup. And indeed, using that approach ground
the emulator to a near halt, even on a computer that should be fairly
overspecced. But bear in mind that I'm a total multithreading n00b and
was only using std::thread, so maybe a multithreading whiz could come up
with an approach that's at least on par with the single threaded version.