Add Discord Rich Presence support
This adds Rich Presence support for melonDS.
This allows anyone with Discord installed on their computer to show which game they're playing.
It additionally allows to show how long the game has been played, as well as showing an image of the game fetched from GameTDB, if available.

On top of this, the PR features some small fixes.